ComRV: Error taking address of overlay func/data with offset != 0
This was previously an assertion that would trigger when referencing
any GlobalValue with a non-zero Offset, even if it the global value
wasn't marked overlaycall or as overlay data.

The assertion has been replaced with a fatal error because it
can be triggered from user code but cannot be handled gracefully.
The fatal error now only occurs if the GlobalValue is marked as
overlaycall or as overlay data.